fix(engine): zero the decided config tree-wide - #946 part 2 close-out
The exit-proof scan (205 TUs, -header-filter, deduped by file/line/column/
check) reported 100 findings still standing after part 1: the non-declined
readability-*/modernize-* remainder never had an owning wave, headers were
never in any wave's count, and six were fallout from part 1's own const-ref
change (the submit lambdas still took result by value once handle_result
stopped consuming it).

By family: nested conditional operators (14) restructured into if/else or
shared helpers (pm_description_text, leaf_text); implicit widening (12) in
constants.hpp given typed first factors (size_t{ 64 } * 1024, the #980
spelling); math parentheses (11); redundant member inits (10) and
use-default-member-init (5) reconciled so a default lives at the member,
once; use-std-min-max (8); DeMorgan simplifications (4); redundant casts
(5); the submit lambdas to const& (6); pass-by-value-and-move (2);
assignment-in-if in url_decode restructured (2); thread_pool's std::bind
replaced with a pack-capture lambda, which also retires the std::result_of
deprecation it instantiated; and the mechanical singles (redundant
declarations, c_str, string init, container-size-empty, parameter-name
mismatch, member-function-const, suspicious-call-argument,
pro-type-member-init in recovery.hpp).

Three suppressions, each with its reason at the site: the version and
platform macros stay macros (build.py's bump regexes and #if consumers are
their contract), oauth_client's {} member inits stay (they silence
-Wmissing-field-initializers at eight aggregate sites), and
EventLoopWorker's field order stays lifecycle-grouped rather than
padding-optimal (one instance per worker thread).

+++ engine/src/db/database.cpp
@@ -1897,8 +1897,9 @@ void Database::prune_runs (int max_runs, int max_age_days) {
         std::chrono::system_clock::now ().time_since_epoch ())
                             .count ();
         // 0 disables the age cap; guard the multiply against overflow.
-        const int64_t age_cutoff =
-        max_age_days > 0 ? now - static_cast<int64_t> (max_age_days) * 86'400'000LL : 0;
+        const int64_t age_cutoff = max_age_days > 0 ?
+        now - (static_cast<int64_t> (max_age_days) * 86'400'000LL) :
+        0;
 
         int kept = 0;
         for (const auto& run : runs) {
@@ -2056,7 +2057,7 @@ int Database::add_inbox_request (const InboxRequest& capture, int64_t max_captur
     int assigned_id = 0;
     retry_on_busy ("append inbox capture", 5, std::chrono::milliseconds (100), [&] {
         impl_->storage.transaction ([&] {
-            assigned_id = static_cast<int> (impl_->storage.insert (capture));
+            assigned_id = impl_->storage.insert (capture);
 
             if (max_captures > 0) {
                 const int64_t stored = impl_->storage.count<InboxRequest> (
@@ -2142,7 +2143,7 @@ const std::vector<PendingResultBody>& bodies) {
             std::vector<int> result_ids;
             result_ids.reserve (results.size ());
             for (const auto& result : results) {
-                result_ids.push_back (static_cast<int> (impl_->storage.insert (result)));
+                result_ids.push_back (impl_->storage.insert (result));
             }
 
             // Dedup within the run: identical bodies (the norm for a load test)
@@ -2164,7 +2165,7 @@ const std::vector<PendingResultBody>& bodies) {
                         blob.run_id  = results[pending.result_index].run_id;
                         blob.hash    = pending.body_hash;
                         blob.content = pending.body;
-                        blob_id = static_cast<int> (impl_->storage.insert (blob));
+                        blob_id      = impl_->storage.insert (blob);
                         blob_ids.emplace (pending.body_hash, blob_id);
                     }
                 }
